added navbar and signing up and login function to landing page inside of navbar
