+++ src/Components/Navbar/navbar.module.scss
@@ -1,0 +1,63 @@
+@import "../../Styling/variables.module";
+@import "../../Styling/mixins.module";
+@import "../../Styling/fonts.module";
+@import "../../Styling/reset.module";
+
+// ? Universal styling
+
+.navbar-container {
+    z-index: 6000;
+    overflow: hidden;
+    // background-color: $cream;
+    background-color: rgba(242, 235, 220, 0.8);
+    position: fixed;
+    bottom: 0vh;
+    width: 100vw;
+    height: 10vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    // margin-top: 50%;
+}
+
+.icon-container {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: nowrap;
+    justify-content: space-between;
+    // background-color: pink;
+    width: 80vw;
+}
+
+.icon {
+    font-size: 2rem;
+    color: $main-font-colour;
+}
+
+// ? Media Queries
+// Based on Bootstrap breakpoints
+
+// * X-Small = Less than 576px - small mobile devices
+@include x-small {
+}
+
+// * Small = Less than 768px - mobile devices
+@include small {
+}
+
+// * Medium = Less than 992px - tablets
+@include medium {
+}
+
+// * Large = Less than 1200px - small desktops
+@include large {
+}
+
+// * Extra Large = Less than 1399px - medium desktops
+@include extra-large {
+}
+
+// * Extra EXTRA Large = More than 1400px - medium desktops
+@include extra-extra-large {
+}
